+++ productservice/src/main/java/com/seckill/productservice/service/impl/LoanProductService.java
@@ -3,50 +3,70 @@ package com.seckill.productservice.service.impl;
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.seckill.productservice.dao.LoanProductDao;
 import com.seckill.common.entity.product.LoanProductEntity;
+import com.seckill.productservice.exception.*;
 import com.seckill.productservice.service.ILoanProductService;
 import org.apache.commons.lang.StringUtils;
-import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.redis.core.RedisTemplate;
+import org.springframework.data.redis.core.ValueOperations;
 import org.springframework.stereotype.Service;
 
+import javax.annotation.Resource;
 import java.util.List;
 
 @Service
 public class LoanProductService implements ILoanProductService {
-    private final LoanProductDao loanProductDao;
+    @Resource
+    private LoanProductDao loanProductDao;
 
-    @Autowired
-    public LoanProductService(LoanProductDao loanProductDao) {
-        this.loanProductDao = loanProductDao;
-    }
+    @Resource
+    private RedisTemplate<String, Object> redis;
 
     @Override
-    public boolean addLoanProduct(LoanProductEntity loanProductEntity) {
-        int i = loanProductDao.insert(loanProductEntity);
-        if(i == 0){
-            return false;
-        }else{
-            return true;
+    public void addLoanProduct(LoanProductEntity loanProductEntity) throws Exception{
+        // redis插入  产品ID：库存
+        ValueOperations<String, Object> opsForValue = redis.opsForValue();
+
+        QueryWrapper<LoanProductEntity> queryWrapper = new QueryWrapper<>();
+        queryWrapper.eq("loan_product_name",loanProductEntity.getLoanProductName());
+        List<LoanProductEntity> i = loanProductDao.selectList(queryWrapper);
+        if(i == null){
+            int insert = loanProductDao.insert(loanProductEntity);
+            if (insert == 0){
+                throw new AddProductionException("添加产品失败");
+            }else{
+                Integer count = loanProductEntity.getStock();
+                opsForValue.set(loanProductEntity.getLoanProductId(),count);
+            }
+        }else {
+            throw new AlreadyExistsProductException("产品已存在，无需重复添加");
         }
     }
 
     @Override
-    public boolean deleteLoanProduct(long loanProductId) {
-        int i = loanProductDao.deleteById(loanProductId);
-        if(i == 0){
-            return false;
+    public void deleteLoanProduct(long loanProductId) throws Exception{
+        LoanProductEntity re = loanProductDao.selectById(loanProductId);
+        if(re != null){
+            int delete = loanProductDao.deleteById(loanProductId);
+            if(delete == 0){
+                throw new DeleteProductionException("删除产品失败");
+            }
         }else{
-            return true;
+            throw new NotFindProductException("找不到指定产品");
         }
     }
 
     @Override
-    public boolean updateLoanProduct(LoanProductEntity loanProductEntity) {
-        int i = loanProductDao.updateById(loanProductEntity);
-        if(i == 0){
-            return false;
+    public void updateLoanProduct(LoanProductEntity loanProductEntity) throws Exception{
+        LoanProductEntity re = loanProductDao.selectById(loanProductEntity.getLoanProductId());
+        if(re != null){
+            int delete = loanProductDao.updateById(loanProductEntity);
+            if(delete == 0){
+                throw new UpdateProductionException("更新产品失败");
+            }
         }else{
-            return true;
+            throw new NotFindProductException("找不到指定产品");
         }
+
     }
 
     @Override
